+++ sopa/segmentation/__init__.py
@@ -1,4 +1,4 @@
 from . import shapes, aggregate, methods, stainings
-from ..patches import Patches2D, BaysorPatches
+from ..patches import Patches2D
 from .aggregate import Aggregator, overlay_segmentation
 from .stainings import StainingSegmentation