+++ CHANGELOG.md
@@ -1,4 +1,16 @@
-## [1.0.14] - 2024-04-22
+## [1.1.0] - 2024-06-11
+
+First post-publication release
+
+### Changed
+- Using `rioxarray` as a default backend for MERSCOPE data if installed
+- Lower RAM usage for channels aggregation
+- Transcript-segmentation API more general (not Baysor-specific)
+
+### Fixed
+- Encoding issue while writing the report (#64)
+
+## [1.0.14] - 2024-04-25
 
 ### Changed
 - Renamed `embed_wsi_patches` to `infer_wsi_patches`